Moved the StoringObserver to the test.helpers
This makes it possible again to make the example package private
