--- net.powermatcher.examples/src/net/powermatcher/examples/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
