+++ net.powermatcher.test.helpers/bnd.bnd
@@ -4,3 +4,4 @@ Bundle-Version: 2.0.0.${qualifier}
 Bundle-Description: A library for testing purposes, that should be statically linked. \
 	Contains many small utilities. This bundle should not be installed in a runtime framework, it is for testing only.
 Export-Package: net.powermatcher.mock, net.powermatcher.test.helpers
+Service-Component: *
