+++ net.powermatcher.test.osgi/src/net/powermatcher/test/osgi/RemoteClusterTests.java
@@ -9,8 +9,8 @@ import java.util.List;
 import net.powermatcher.api.messages.PriceUpdate;
 import net.powermatcher.api.monitoring.events.IncomingPriceUpdateEvent;
 import net.powermatcher.api.monitoring.events.OutgoingBidUpdateEvent;
-import net.powermatcher.examples.StoringObserver;
 import net.powermatcher.test.helpers.PropertiesBuilder;
+import net.powermatcher.test.helpers.StoringObserver;
 
 import org.osgi.service.cm.Configuration;
